Add files via upload
In this folder, the output of the IBM-PD for the simulation of 1-year ideal rearing cycle is presented. Due to maximum upload size, the complete file was seperated into these files, in which each of them contains the weight trajectory simulated by the model of 500 individuals. Mortality has already been included